[#13] refactor: rename package com.ktomek.yamv.viewmodel to com.ktomek.yamv.retainer
